requirements.txt
asgiref==3.5.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
daphne==4.0.0
Django==4.1.4
django-cors-headers==3.14.0
django-generic-filters==2.1.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
munch==2.5.0
openapi-codec==1.3.2
packaging==21.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.1.1
pyparsing==3.0.9
pytz==2022.6
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
simplejson==3.18.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
zope.interface==6.0
